add getExtraFilms, templates, if for 1 popup
