+++ src/main.js
@@ -1,11 +1,10 @@
-import {CARD_EXTRA_COUNT} from './const.js';
 import {CARD_COUNT} from './mock/const.js';
 import Rank from './view/rank.js';
 import SiteMenuView from './view/site-menu.js';
 import FilmsContainer from './view/films-container.js';
 import FilmCard from './view/film-card.js';
 import ShowMore from './view/show-more.js';
-import Comment from './view/comment-template.js';
+import Comment from './view/comment.js';
 import {generateFilm} from './mock/film.js';
 import {generateUser} from './mock/user.js';
 import Popup from './view/popup.js';
@@ -16,35 +15,13 @@ import FilmListCommented from './view/list-commented.js';
 // import Stat from './view/stat.js';
 import Sort from './view/sort.js';
 import {render, remove, RenderPosition} from './utils/render.js';
-import {onEscKeyDown} from './utils/common.js';
+import {onEscKeyDown, getExtraFilms} from './utils/common.js';
 
 const films = new Array(CARD_COUNT).fill().map(generateFilm);
 
-const topFilms = films
-  .slice()
-  .sort((a, b) => {
-    if (a.rating > b.rating) {
-      return -1;
-    }
-    if (a.rating < b.rating) {
-      return 1;
-    }
-    return 0;
-  })
-  .slice(0, CARD_EXTRA_COUNT);
+const topFilms = getExtraFilms(films, `rating`);
 
-const popularFilms = films
-  .slice()
-  .sort((a, b) => {
-    if (a.comments.length > b.comments.length) {
-      return -1;
-    }
-    if (a.comments.length < b.comments.length) {
-      return 1;
-    }
-    return 0;
-  })
-  .slice(0, CARD_EXTRA_COUNT);
+const popularFilms = getExtraFilms(films, `comments`);
 
 const user = generateUser();
