+++ src/view/film-container-template.js
@@ -1,0 +1,3 @@
+export const createFilmsTemplate = () => {
+  return `<section class="films"></section>`;
+};
