+++ src/view/film-number-template.js
@@ -1,0 +1,3 @@
+export const createFilmNumberTemplate = (num) => {
+  return `<p>${num ? num : `0`} movies inside</p>`;
+};
