+++ src/view/list-commented-template.js
@@ -1,0 +1,6 @@
+export const createfilmListCommentedTemplate = () => {
+  return `<section class="films-list films-list--extra">
+    <h2 class="films-list__title">Most commented</h2>
+    <div class="films-list__container"></div>
+  </section>`;
+};
