+++ src/view/list.js
@@ -1,11 +1,5 @@
-import AbstractView from "./abstract.js";
-
-const createfilmListTemplate = (exist) => {
-  return `<section class="films-list">
-    <h2 class="films-list__title">${exist ? `All movies. Upcoming` : `There are no movies in our database`}</h2>
-    <div class="films-list__container"></div>
-  </section>`;
-};
+import {createfilmListTemplate} from './list-template.js';
+import AbstractView from './abstract.js';
 
 export default class FilmList extends AbstractView {
   constructor(exist) {
