+++ src/view/show-more-template.js
@@ -1,0 +1,3 @@
+export const createShowMoreTemplate = () => {
+  return `<button class="films-list__show-more">Show more</button>`;
+};
