+++ src/view/sort.js
@@ -1,12 +1,5 @@
-import AbstractView from "./abstract.js";
-
-const createSortTemplate = () => {
-  return `<ul class="sort">
-    <li><a href="#" class="sort__button">Sort by default</a></li>
-    <li><a href="#" class="sort__button">Sort by date</a></li>
-    <li><a href="#" class="sort__button sort__button--active">Sort by rating</a></li>
-  </ul>`;
-};
+import {createSortTemplate} from './sort-template.js';
+import AbstractView from './abstract.js';
 
 export default class Sort extends AbstractView {
   getTemplate() {
